+++ src/cli.rs
@@ -3,16 +3,25 @@
  * License: MIT
  */
 
-use clap::Parser;
+use clap::{Parser, Subcommand};
 use std::path::PathBuf;
 
 /// Command-line arguments
 #[derive(Parser, Debug)]
 #[clap(about, author, version)]
 pub(crate) struct Cli {
-    /// Specify configuration filename (e.g. `config.toml`)
-    #[clap(short = 'c', long = "config")]
-    pub config_filename: PathBuf,
+    #[command(subcommand)]
+    pub command: Command,
+}
+
+#[derive(Debug, Subcommand)]
+pub(crate) enum Command {
+    /// Run the Verbleiber client
+    Run {
+        /// Specify configuration filename (e.g. `config.toml`)
+        #[clap(short = 'c', long = "config")]
+        config_filename: PathBuf,
+    },
 }
 
 pub(crate) fn parse_cli() -> Cli {